Address CECE warnings from clang-tidy

+++ src/io/cece_data_ingestor.cpp
@@ -114,7 +114,7 @@ void CeceDataIngestor::SetField(const std::string& name, const double* data, int
         }
     } else {
         // For 3D data, use direct memory copy
-        std::memcpy(host_view.data(), data, n_lev * n_elem * sizeof(double));
+        std::memcpy(host_view.data(), data, static_cast<size_t>(n_lev) * static_cast<size_t>(n_elem) * sizeof(double));
     }
 
     // Allocate device view if it doesn't exist or has wrong shape
